remove unneeded expensive assert
