--- py-polars/tests/test_fallible.py
@@ -1,12 +1,0 @@
-import io
-
-import pytest
-
-import polars as pl
-
-
-def test_not_found_error() -> None:
-    csv = "a,b,c\n2,1,1"
-    df = pl.read_csv(io.StringIO(csv))
-    with pytest.raises(pl.NotFoundError):
-        df.select("d")
